requirements/requirements.txt
fastapi==0.128.7
starlette>=0.50.0
uvicorn[standard]>=0.39
pydantic>=2.10
sqlalchemy
psycopg[binary]>=3.3
python-decouple
tzdata
